DBOX #36 Isolate logging container infrastructure, add thorough logging logic inside Django apps

+++ backend/live/schema.py
@@ -1,10 +1,14 @@
 import graphene
+import logging
 from graphene_django import DjangoObjectType
 from graphql import GraphQLError
 from django.utils import timezone
 from datetime import datetime, timedelta
 from .models import Scheduling
 from content.models import Publication
+from users.utils import get_client_ip
+
+logger = logging.getLogger('django.live')
 
 class SchedulingType(DjangoObjectType):
     class Meta:
@@ -28,9 +32,8 @@ class CreateScheduling(graphene.Mutation):
 
     def mutate(root, info, publication_id, time):
         user = info.context.user
-        if not user.is_authenticated:
-            raise GraphQLError("You must be logged in as a privileged user to add to the schedule")
-        if not user.has_perm("live.add_scheduling"):
+        if not (user.is_authenticated and user.has_perm("live.add_scheduling")):
+            logger.warning(f"Unauthorized Scheduling creation attempt from {user if user.is_authenticated else "visitor"}@{get_client_ip(info.context)}")
             raise GraphQLError("You do not have the required permissions to add to the schedule")
         # WARNING : a timedelta of one day should be added in production : if time < (datetime.now() + timedelta(days=1))
         if time < datetime.now():
@@ -39,6 +42,7 @@ class CreateScheduling(graphene.Mutation):
             publication = Publication.objects.get(id=publication_id)
         except Publication.DoesNotExist:
             raise GraphQLError("This Publication does not exist")
+        logger.info(f"{publication} successfully scheduled at {time} by {user}")
         scheduling = Scheduling(publication=publication, time=time)
         scheduling.save()
         return CreateScheduling(scheduling=scheduling)
@@ -51,9 +55,8 @@ class DeleteScheduling(graphene.Mutation):
 
     def mutate(root, info, scheduling_id):
         user = info.context.user
-        if not user.is_authenticated:
-            raise GraphQLError("You must be logged in as a privileged user to delete from the schedule")
-        if not user.has_perm("live.delete_scheduling"):
+        if not (user.is_authenticated and user.has_perm("live.delete_scheduling")):
+            logger.warning(f"Unauthorized Scheduling deletion attempt from {user if user.is_authenticated else "visitor"}@{get_client_ip(info.context)}")
             raise GraphQLError("You do not have the required permissions to delete from the schedule")
         try:
             scheduling = Scheduling.objects.get(id=scheduling_id)
@@ -61,6 +64,7 @@ class DeleteScheduling(graphene.Mutation):
             raise GraphQLError("This Scheduling does not exist")
         if scheduling.time < timezone.now():
             raise GraphQLError("The schedule for this date and time has already been fixed. You cannot update it")
+        logger.info(f"Scheduling {scheduling.id} ({scheduling}) successfully deleted by {user}")
         scheduling.delete()
         return DeleteScheduling(success=True)
